# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Truskill Games",
        "description": "API for the Truskill Games",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user.",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "The new User email (unique)",
                                        "type": "string",
                                        "example": "example@test.com"
                                    },
                                    "name": {
                                        "description": "The new User name (unique, max:255)",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "password": {
                                        "description": "The new User password (min:8)",
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "description": "The password confirmation (same as 'password')",
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserWithToken"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to create a new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR_CREATE_FAILED"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Provided data is not valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR500"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login the user.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "The User email",
                                        "type": "string",
                                        "example": "example@test.com"
                                    },
                                    "password": {
                                        "description": "The User password",
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserWithToken"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR_CREATE_FAILED"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "the user not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR_NOT_FOUND"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Provided data is not valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR500"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the user.",
                "operationId": "logout",
                "responses": {
                    "201": {
                        "description": "User logged out."
                    },
                    "400": {
                        "description": "Failed to create a new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR_CREATE_FAILED"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token expired or not provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR401"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ERR500"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ERR401": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ERR422": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The field_name field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ERR500": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Server Error"
                    }
                },
                "type": "object"
            },
...
            "ERR_CREATE_FAILED": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": "600"
                    },
                    "error": {
                        "type": "string",
                        "example": "Failed to create a resource."
                    },
                    "details": {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "More detailed error message."
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ERR_NOT_FOUND": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": "604"
                    },
                    "error": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
...
            "UserWithToken": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 9876
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "example@test.com"
                    },
                    "token": {
                        "description": "User api token",
                        "type": "string",
                        "example": ""
                    },
                    "createdAt": {
                        "description": "User creation datetime",
                        "type": "string",
                        "example": "2022-04-20 13:15:31"
                    },
                    "updatedAt": {
                        "description": "User last modification datetime",
                        "type": "string",
                        "example": "2022-04-24 18:00:00"
                    }
                },
                "type": "object"
            }
        },
...
    },
    "security": [
        []
    ]
}